apis/fastapi/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
fastapi==0.115.8
greenlet==3.1.1
idna==3.10
pydantic==2.10.6
pydantic_core==2.27.2
python-dotenv==1.0.1
sniffio==1.3.1
SQLAlchemy==2.0.38
starlette==0.45.3
typing_extensions==4.12.2
